RFC: ARM64: Split arm64 codegen into scalar and vector (SVE and NEON).
This is a first CL in the series of introducing arm64 SVE support
in ART. The patch splits the codegen functionality into scalar and
vector ones and for the latter introduces NEON and SVE
implementations. SVE one currently is an exact copy of NEON one -
for the sake of testing and an easy diff when the next CL comes
with an actual SVE instructions support.

The patch effectively doesn't change any behavior; NEON mode is
used for vector instructions, tests pass.

Test: test-art-target.
Change-Id: I5f7f2c8218330998e5a733a56f42473526cd58e6
Signed-off-by: DennySPb <dennyspb@gmail.com>