Use side vector while sweeping jvmti-weak-table
In the code we were erasing all <K,V> pairs for which object reference
(Key) is changed, and then emplacing the ones which were moved. This
is not compatible with in-place compaction algorithms as then there is a
possibility that a to-space object reference being added to the map is
already there as a from-space reference.

In this CL, we change it to use a side vector to hold all updated <K,V>
until the end of map traversal, and then insert all of them back in the
map.

Bug: 160737021
Test: ART_USE_READ_BARRIER=false art/test/testrunner/testrunner.py
--host --debug -t 905-object-free

Change-Id: Ib416ca0b21d4a9fc44721495b798f82be9387900
Signed-off-by: minaripenguin <minaripenguin@users.noreply.github.com>