Check for deoptimized method only when there are method entry callbacks
We always update entrypoint of the method to point to InterpreterBridge
if the method is deoptimized. So, we don't need to check if the method
is deoptimized when there are no method entry callbacks. Method entry
callbacks could set new breakpoints and hence we need to check if the
metnod is deoptimized after running any method entry listeners.

Bug: 253232638
Test: art/test.py
Change-Id: Iadb25a96307646ad7119162d56038729930767cb
Signed-off-by: Pranav Vashi <neobuddy89@gmail.com>