Speed up graph checker for graphs with instructions with many uses
As part of VisitInstruction, we verify that the instruction's
input_record is present in the corresponding input's GetUses. If
an instruction is used in many places (e.g. 200K+ uses), the
linear search through GetUses is too slow. We can use bookkeeping
to search in a set, instead of a list.

This makes dex2oatd over an order of magnitude faster for such cases.

Bug: 301264418
Test: Manually compile the app+profile in the bug
Change-Id: I3dd6d1b9b2fd65a4bf5ee8324602a412e3f1715f
Signed-off-by: Pranav Vashi <neobuddy89@gmail.com>